delete button colour changed from red to white. Buffer read updated to cover longer files. Poll modified to get called if request doesn't contain the whole body

+++ inc/Client.hpp
@@ -8,13 +8,13 @@ class Client
 private:
 	int fd; // The client-specific socket file descriptor, returned by accept(), in the Socket class.
 	// Used for reading(receiving requests) and writing(sending responses).
-	int request_size;
+	long request_size;
 	std::string raw_request; // Stores the full HTTP request as a raw string
 	
 public:
 	Client() {} ; // Default constructor. Not used in this context, but defined for completeness.
 	Client(Socket& socket); // Calls acceptClient() from the passed-in Socket object.
-	Client(int poll_fd, int requset_size);
+	Client(int poll_fd, long requset_size);
 	~Client(); // Automatically close the socket when the client is destroyed.
 
 	int getClientFd() const; // Simple getter for the socket file descriptor.
